ngan/requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.8.0
attr==0.3.2
attrs==25.1.0
audioread==3.0.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
DateTime==5.5
decorator==5.1.1
deep-translator==1.11.4
duckduckgo_search==7.3.2
emoji==2.14.1
ephem==4.2
fake-useragent==2.0.3
Flask==3.1.0
fonttools==4.56.0
forex-python==1.8
frozenlist==1.5.0
google==3.0.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-vision==3.9.0
google_search_results==2.4.2
googleapis-common-protos==1.66.0
googletrans==4.0.2
grpcio==1.70.0
grpcio-status==1.70.0
gTTS==2.5.4
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
langdetect==1.0.9
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
LunarCalendar==0.0.9
lxml==5.3.0
lyricsgenius==3.2.0
MarkupSafe==3.0.2
matplotlib==3.10.0
msgpack==1.1.0
multidict==6.1.0
munch==4.0.0
numba==0.61.0
numpy==2.1.3
opencv-contrib-python==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
primp==0.12.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
py-cpuinfo==9.0.0
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.19.1
pyjokes==0.8.3
pyparsing==3.2.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.1
pyzbar==0.1.9
regex==2024.11.6
rel==0.4.9.19
removebg==0.4
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.29.0
setuptools==75.8.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
speedtest-cli==2.1.3
Telethon==1.38.1
threadpoolctl==3.5.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wikipedia==1.4.0
wsproto==1.2.0
yarl==1.18.3
youtube-search==2.1.2
youtube-search-python==1.6.6
yt-dlp==2025.1.26
zlapi==1.0.3
zope.interface==7.2
unidecode==1.3.6



